pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "chanterelle"
version = "0.1.0"
description = "Analytics Agent Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "pyarrow>=17.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "boto3>=1.35",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]

[tool.setuptools.packages.find]
include = ["shared*", "services*"]

[tool.ruff]
line-length = 100
target-version = "py311"
